Playing around with ordering.
This is still not usable in its current
iteration. However, the ordering provides
some potential ideas of how to structure the
data frame. The big issue is that labels
between models may be the same but could
theoretically have completely different
levels. How do we proceed with the match
then?