Add bugfixes, basic necessary options to output_latex.
caption, group_names, repeat_heading, and bottom_rule are now all options. Note that we still have env as an option but we may want to 'fix' it to longtable (and dcolumn, as it is right now), as much of the overall appearance depends on it. For the same reason, we will probably consider booktabs mandatory.