Add reg_add_labels (WIP?).
This implementation of of reg_add_labels is
very complicated due to:

1. Interactions with * needing to be parsed
   into separate components.
2. Factor levels having label information
   when levels are a mix of factors and
   non-factors.

The transformations in this code are not
ideal and they call into question a need
for perhaps a bigger refactor, in
particular how labels and factor levels are
'stored'. However, the first step is to
plug in the missing functionality and then
perhaps we shall address the brittleness of
these transformations.