Fix and streamline reg_add_labels.
The previous iteration was not only brittle,
but the fact an extra column was added to
the df also broke reg_bottom_se. In
addition, there was an error in the label ->
factor fill. This is fixed and simplified
in the new version, but there is an
implicit expectation that the label /
factor multipliers are the same, which
sounds right but potentially should not
always hold true. The whole unite call of
course points to the fact that storing
labels / factor levels of interactions in
one column -- with a brittle string no
less -- is far from ideal. We may need to
revisit these assumptions in the future to
make the implementation more robust.